ai-service/requirements.txt
kafka-python==2.0.2
python-json-logger
# S3 추가.